requirements.txt
requests
Pillow
matplotlib
numpy
tqdm
